Add files via upload
I'm not sure why I keep getting a syntax error at line 25. I attempted to just run it as:

function getResults(query){
  (fetch('https://dataservice.accuweather.com/locations/v1/search?apikey=${apiKey}&q=${query}')
    .then(location => {
      return location.json();
    }).then(updateCity);
}

But kept getting CORS errors, and the way its done in the new .js file seemed to be a workaround? But I'm not totally sure if its valid or where to go next.